Prepare for Pull Request
Features
=======
- Integration of mixlib-log for application logging
- Creation of a searchable mixin that provides scoped JQL and pagination
- Support for ruby block on the _all_ idiom
- Support for casting has_many and has_one relationships as primitive
ruby types (String, Fixnum, Array, etc)
- Support for nesting api calls under other nodes (i.e.
/rest/api/2/issue/{issueIdOrKey}/worklog) where work log is nested
under issue.
- Support for the Jira expanding API. Some attributes need to be
expanded to see them, by default attributes will not be expanded,
however when an an attribute the requires expanding is needed, the
object will be reloaded with all expanding attributes expanded.
- Added has_one relationship on JIRA::Resource::User for fields that
may not be present for some disabled users.

Testing
======
- Updated old tests to pass given the addition of new features
- Added tests to exercise search and paging features.
- Added tests to exercise the expanding API
- Added tests to exercise ruby blocks with the _all_ idiom
- Added tests for labels on issues
- Added intégration tests for changlog
- Added resource tests for change log
- Added mock data

Notes
=====
Some features where not explicitly tested because they are exercised in
the normal course of testing. i.e. Most of the changes to base.rb