Very basic support for transitions.
Fixes #24

Transitions are issue dependent, so we can't just load all transitions, neither are they tied to the issue via a key which makes fetching the available transitions a bit, messy.

Anyhow, this allows all the transitions associated with an issue to be fetched, and then a transition can be told to run against an issue_key.

transitions = client.Transition.forIssue issue_key
transitions.each { |transition|
    # If this is the transition we want to run…
    transition.transitionIssue issue_key
}