handle non json error response bodies
When making a request with the http_client and the request fails to authenticate (eg expired JIRA password), JIRA responds with a 401 Unauthenticated error. We rescue this exception, and then try to parse the body into our attrs. However, the body that is returned in this state is invalid json, so that throws an exception in our exception handler, which we previously swallowed.

This change simply handles that by falling back to reading some info off the original exception (Net::HTTPUnauthorized -> Net::HTTPClientError -> Net::HTTPResponse). Specifically the class name, response code and message.