Bug updating the resource after a save
If the attribute to update is a hash and it is nill, then the wrong
attributes are being updated.

I found this bug while I was updating the assignee in an issue without
assignee. In that case, with this code:

issue.save!('fields' => { 'assignee' => user.attrs }) if issue.assignee.nil?

in this case, all the attributes of the user are being set to the issue,
for instance the issue.id becomes user.id, etc ...

This patch solves the issue and pases all the tests.