HttpClient: Allow specifying ca_file
I'm working in an environment where we have our own certificate
authority, and I need to pass a CA file. It doesn't look like there's
currently a way to do so.

I see a similar patch was submitted in
https://github.com/sumoheavy/jira-ruby/pull/279 and it was just
withdrawn because the submitter didn't need it anymore. Now that I'm
seeing a need, I wonder whether there is a possibility of adding it.

I see a monkey-patch solution being offered in
https://github.com/sumoheavy/jira-ruby/issues/140#issuecomment-196534011
but my colleagues are leery of monkey-patching.

I also see from my searches that perhaps setting the environment
variable SSL_CERT_FILE may be able to achieve this, but we're also a
little leery of an implicit reliance on the environment variable, and
think it's better to be able to explicitly specify the ca file.